Disable metadata edit button when the topic is dirty
This prevents users from adding content, then trying to edit
the metadata, saving that -- which marks the topic as clean again,
even though it is not (see #218). This is a temporary fix. The
permanent fix would be to keep the isDirty state when the metadata
changes, but not when the content gets updated.